Return to client macros for tgsay (#6320)
# About the pull request

This PR is a follow up to #6244 effectively reverting the changes ported
from https://github.com/tgstation/tgstation/pull/75463

Instead of using the down() proc on the server, we are once again doing
the tgsay open via winset on hotkey so it is more responsive. However,
to fix the issue with tgsay opening when say another window is focused,
it now has an embedded winget for map.focus and thats sent in the
payload.

# Explain why it's good for the game

Should make tgsay open more responsively again.

This was the reason why I needed to make the change in the first place: 

![bug](https://github.com/cmss13-devs/cmss13/assets/76988376/0288d3d1-e3c9-4cf7-95c2-9f7ef88095da)

# Testing Photographs and Procedure
But now that bug is fixed, and we're opening the old way again!


![tgsay](https://github.com/cmss13-devs/cmss13/assets/76988376/0997f797-7b99-4e6e-acda-722a783b010f)

# Changelog
:cl: Drathek
ui: TGSay is now opened the same way it used to for more responsiveness
/:cl: