Adds a sensor tower to Shivas Snowball (#6309)
# About the pull request

If a competent mapper wants to fix this, please re-do it :)

I've placed it on the xeno-side, between Med-Sec and the Labs, might be
too hard to get too, idk. Don't want to favor north/south, to prevent a
Solaris situation where Lambda is basically mandatory to hold.

# Explain why it's good for the game

Every other map has a sensor tower, Shivas was temporarily removed when
they got added.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/cmss13-devs/cmss13/assets/5618080/3d6b1357-c23c-418b-95e2-fb2fddcbc8ab)


</details>


# Changelog
:cl:
add: Adds a sensor tower to the far East of Shivas, between Med-Sec and
the Research labs
/:cl: