Fixes some improper APC placements on Shivas Snowball (#6315)
# About the pull request

Fixes a few APCs that were previously placed on top of indestructible
dirt, preventing the APC terminal being accessible, and thus making them
impossible to move or repair.

Also changes the area path used by the small North generator
room/communications room. This room previously shared the same area
pathing as the room on the West side of LZ2, both of which had APCs.
Doubling up APCs in one area is a mapping sin and makes resolving any
issues with either of the APCs extremely frustrating.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Bad object placement is bad (the design is very human).

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>
This is the room with the area pathing changes:


![image](https://github.com/cmss13-devs/cmss13/assets/80382633/e52a54f8-98af-49bc-8d96-8386c952eebb)

See MapDiffBot for the rest, I don't think it's particularly practical
for me to screenshot multiple 1 tile changes.
</details>


# Changelog
:cl:
fix: Fixes improperly placed APCs on Shivas Snowball
/:cl: