AMR Anti-Focusing nerfs (#6311)
# About the pull request
Makes the AMR counter reset to 0 after two fully charged shots
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Being rewarded for focusing down a single xeno is toxic and bad for the
game. Instead, after hitting it with 2 heavy shots you will have a
choice between continuing to focus that xeno or focusing a different
xeno with no loss of efficiency.
According to the guy who added AMR focusing only one target is terribly
inefficient so this is infact a buff.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: AMR will now reset its focus after two fully charged aimed
shots on the same target.
/:cl: