Synthetic, Compact-Nailgun Holster (#6277)
# About the pull request

Adds a customized compact nailgun holster as a item vended alongside the
compact nailgun in the Synthetic vendor.

The item functions akin to a regular pistol holster, however it can only
hold a compact nailgun and two nailgun magazines. It can not hold the
regular nailgun or any other pistol or pistol magazine.

# Explain why it's good for the game

Additional options for inventory customization for Synthetics units in
storing the compact nailgun. Arguably this item is balanced as the
compact nailgun already is easy to store in multiple alternative storage
options, this simply adds a more streamlined storage area, and befits
the nature of a Synthetic to go the extra step and setup a customized
holster.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

![Screenshot 2024-05-11 20 34
10](https://github.com/cmss13-devs/cmss13/assets/6595389/926a6b9f-0645-4987-9acf-f86b410104fc)


</details>


# Changelog
:cl:
add: The Compact Nailgun that the USCM Synthetic can vend now comes with
a unique nailgun holster, capable of holding the nailgun and two spare
magazines.
/:cl:

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>