Vehicle projectile origin adjustment (#6245)
# About the pull request

Vehicle projectiles now start on the vehicle and travel outward, without
hitting the vehicle. As a side effect, `Adjacent()` checks now accept a
list of objects to ignore.

Origins are adjusted to the point visually closest to the hardpoint's
sprite. For most hardpoints this is the center of the vehicle, with the
only exceptions being the APC weapons (all along different edges) and
the secondary flamer (flame origin is walked until not on the vehicle).

Since projectiles now must travel farther to reach the front of the
vehicle, some shorter ranged ammos are adjusted to compensate.

The hardpoint check `clear_los()` is removed since now all origins are
inside the vehicle's bounds.

Removed redundant call to `display_ammo()` when firing secondary flamer.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

Projectiles originating from the tile in front of a vehicle make it
extremely hard or even impossible to engage targets adjacent to it. With
this change the projectile, and resulting cone of fire, starts where you
would expect it to.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
qol: vehicle weapon projectiles start from the weapon
/:cl: