Project ARES: Almayer Sentries (#6268)
# About the pull request

Changes the sprites for the sentry deployment systems aboard the USS
Almayer to use a custom floor based holder, rather than a floating wall
mount.

Also adds mini sentry deployers, along with an AI-Core subtype that
autodeploys on lockdown.

# Explain why it's good for the game

Currently the sentry deployment system sprites are designed as if
they're mounted to walls, and they usually are, but on the Almayer
they're essentially floating next to the fuel pumps. This changes that.

Also it allows the AI Core to be protected from xenos without subjecting
non-existant joes to melee combat.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/cmss13-devs/cmss13/assets/41653574/a4a671a2-2772-4c13-8f84-c11cf3bd807b)

![image](https://github.com/cmss13-devs/cmss13/assets/41653574/61f83c72-aa40-4a9b-b528-8fd86b4df5f5)


</details>


# Changelog
:cl:
imageadd: added sprites for floor-mounted sentry deployment systems.
add: Added mini-sentry deployment systems, and an AI Core subtype.
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>