Adds UNO cards to the Lifeboat Supply Cabinets (#6326)
# About the pull request

see title

# Explain why it's good for the game

It can get pretty boring for marines on the lifeboat depending on how
long it takes for the round to wrap - this gives them something to do if
they want.

# Testing Photographs and Procedure

<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl:
add: lifeboats now come stocked with uno cards, for all your fun-filled
needs
/:cl:

---------

Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>