Whiskey Outpost for burrower add the ability to make tunnel but remove the ability to burrow. (#6283)
# About the pull request
A-Remove the ability for the burrower to burrower.
B-Add the ability for the burrower to make tunnels.
C-added some more C4 to the marine weapon vendor.
D-updating the clothing vendors
1-remove telephone.
2-updating the vendor to what is in main mode.
3-keep headsets.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
A,B,C the idea is to allow burrower to speed up movement on conquered
ground.(having to restart from the bottom of the map every-time is a
pain for xenos.
the extra C4 is allow engi to deal with the ability of making tunnel.
D the idea is to update the vendor so that it as all the items that have
some usage in WO.

burrowing would be too powerful with this ability so i am removing it.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: for burrower in whiskey outpost add the ability to make tunnel
but remove the ability to burrow.
balance: add C4 to the wo weapon vendor.
balance: updating the clothing surplus vendor on to be as close as main
mode as possible.
/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>