Makes it possible for all Officers and all NCO's (E4+) to recommend people for medals. (#6343)
# About the pull request

I've been meaning to do this.

Corporal and above + all officers can recommend medals.


# Explain why it's good for the game
# Testing Photographs and Procedure
I tested it.

Tried to give a medal recommendation as a PFC: Failed
Tried to give a medal recommendation as a FTL: Success
Tried to give a medal recommendation as an SO: Success
Tried to give a medal recommendation as XO: Success

<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: All officers and NCO's can now recommend people for medals.
/:cl: