Project ARES: Print ASRS Log (#6333)
# About the pull request

Allows the ASRS Audit Log to be printed from the ARES Interface.

# Explain why it's good for the game

Many times I've seen CMB request access to the audit log for the
purposes of checking on black market activity. Whilst this is quite
possible, there's no simple way to take a copy of those records. Now
there is.


# Testing Photographs and Procedure
(This image has a bug in it but that has been fixed)

![image](https://github.com/cmss13-devs/cmss13/assets/41653574/25ad5f12-7c8f-48c8-8798-b2039f5f7d0c)


# Changelog
:cl:
add: Added a print button to the ASRS audit log.
/:cl:

---------

Co-authored-by: Vero <73014819+vero5123@users.noreply.github.com>
Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>