Project ARES: Working Joe uniform cameras. (#6344)
# About the pull request

As title, this adds cameras to Working Joe uniforms, tied to the ARES
Core camera consoles. As a result of this it is no longer possible to
manually build more cameras for the AI Core, or add the network to other
consoles.

# Explain why it's good for the game

Makes sense, at least to me, that the Working Joes would be able to
operate as a camera feed in some regard. This way it ties it to ARES
using them as mobile eyes.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added cameras to Working Joe uniforms.
/:cl: