Fix range usages (#6378)
# About the pull request

So it turns out some usages of range are wrong
https://www.byond.com/docs/ref/#/proc/range with the arguments in the
wrong order. This might have some unintended consequences as some code
that never worked now works...

# Explain why it's good for the game

Fixes issues such as 

![image](https://github.com/cmss13-devs/cmss13/assets/76988376/8c64f46a-7dc8-4bca-9ddf-d621542f0cb4)

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: Fixed some usages of the proc range such as for delimbing
/:cl: