Lots of various refactoring and fixes:
Fixed minor/major defines compared mathematically but as strings.
Removed broken logic to announce a major buff (broken because of above)
Added distant roar sfx to purchasing buffs
Removed/refactored most code handling pylons being in control of a buff
Refactored candidate picker code
Most buffs are now are repurchase able
Opted to use marine announcement instead (as is the lists would be "sticky" for a destroyer. E.g. if you were unconscious when it spawned you'd never get messages for it). And generally cleaned up announcements.
Fixed purchasing buffs not updating the hive status count
Added an evolve time to destroyer caste so it would be announced
Fixed the incorrect timing for hive buffs (added round start time twice)
The same lineage of a buff is now deemed incompatible (e.g. no minor boon of health + major boon of health)
Changed the hatchery layer back to 3 because the issue was plane instead
Fixed the existing destroyer check in hive (misspelled, and you didn't type the path out compeltely so it was pathing to xenomorph because `destoryer` would then be the local var name), and converted it to a locate instead.
Fixed the dictionary list input tgui crash
Added structure density checking and remove the no tunnel area check for hatchery