Requisitions Line Tutorial (#5909)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

Adds a tutorial/simulator for the Requisitions Line. Players are asked
by NPCs to vend them several items. There are 3 training NPCs with
pre-made requests and highlights, after which the tutorial enters an
endless mode for training against increasingly complex requests.

This still has some rough edges, notably that it can ask randomly for
non-obvious names much like in real game such as "Frag" refering to
HEDPs. A new player might need for the NPC to repeat the order with a
different name before getting it right.

Also it might be worth finding a way to do highlighting in the vendor UI
for this...

All items dispensed in the tutorial are cloned as props and will be
missing some overlays and be non interactible, because it's the simplest
way i found so the player doesn't explode everything.

# Explain why it's good for the game
Gives people an opportunity to approach a simple yet potentially
stressful activity. Line rush can be tough when starting the game. This
only contains a tutorial for the line though, other reqs system will
need to be made separately

# Testing Photographs and Procedure
Ran through it a few times testing for unintended interactions

Older video of initial implement:

https://cdn.discordapp.com/attachments/964684928161808384/1178420842347102218/reqline-tutorial.mp4?ex=65f747ee&is=65e4d2ee&hm=797dfdfe2693878f4dc69fb976060a285944023935b2e0b5216670102c37366a&

# Changelog
:cl:
add: Added a Tutorial for tending to the Requisitions Line
/:cl:

---------

Co-authored-by: John Doe <johndoe@gmail.com>