techpoint price change (#6410)
# About the pull request

Halves the price of OB, supply points, and CAS budget. OB gives one
shell, supply budget gives 50 points from 100, CAS budget 2000 from 2500
(FACTCHECK, it is used one per two days( not twice a year) and is not
worth even the 5 points)

# Explain why it's good for the game

Makes it easier to diversify for command and reduces the extre
commitment when buying certain stuff, ten intel points is huge portion
of budget on loads of rounds. Also having to buy two of the same OB
shell is quite the commitment and this makes it so that it costs the
same when you diversify the shells you buy and more if you buy two of
the same. Preventing foxtrot buy buying ds or req budget just is not
done as you have better options aka foxtrot and saving up for nuke.
Dropship budget was never bought and even this buff does not even place
it in line with other stuff, 3 -5 firemissions of budget or 5 more
marines or OB.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: dropship fabricator extra budget price 12 -> 6 , 2500 -> 2000
points
balance: supply extra budget 15 -> 7, 10k -> 5k
balance: all OB 10 -> 5 intel points, one shell and five fuel
/:cl:

Co-authored-by: vincibrv <vojtechvincibr@seznam.cz>