Fix ToD not getting updated if ghostized without a client (#6412)
# About the pull request

This should fix the edge case where a mob is ghostized and they don't
currently have a client. Normally the player_details is accessed via the
client, but we can fallback to the GLOB need be.

# Explain why it's good for the game

Fixes an edge case where a time of death is not recorded for a player as
far as the larva queue is concerned.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Drathek
fix: Fixed larva queue time of death not getting recorded if client is
missing
/:cl: