Personal Civilian Clothing Storage Unit For Marines (#6401)
# About the pull request

Adds a set of civilian clothing vendors on the Almayer in the Gym.,
these vendors can be used by anyone to vend civilian style clothing, the
contents are almost identical to the Synthetic clothing vendor, minus
the military themed clothing and unique backpack/similar options. These
vendors are locked behind a set of shutters with the same high-level
access requirements as the Dress uniform vendors, so CIC can determine
when access is authorized.

Anyone can access the vendor, each person has a 120 point limit to what
they can buy.

# Explain why it's good for the game

Grants regular Marines the ability to vend their own personal civilian
clothing, for events or whenever CIC authorizes it (Maybe a event where
the marines are going on shore leave before something terrible
happened?).

The advantage here is to further allow regular players to craft their
own characters apperance and roleplay what they might wear off-duty, in
circumstances that allow for it.

The area is locked thus CIC has control on when Marines can access it,
and there is also the fact it is the job of the MPs to ensure Marines
are wearing proper clothing.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
mapadd: Civilian Clothing Storage Vendors have been added to the Gym on
the USS Almayer. These allow anyone to vend civilian clothes. These
vendors require CIC access to unlock however.
/:cl:

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>