Normalises the HUD damage overlay so synths and predators HUD overlays trigger at the same %-wise damage as humans. (#6452)
# About the pull request

As human, you get the max damage overlay at 85 damage, i.e 15% health
As synth, you get the max damage overlay also at 85 damage, but that's
43% health.

Normalises this, so synths get it at 15%, and so on.

# Explain why it's good for the game

If you've still got gas left in the tank, there's no reason to overlay a
large and disturbing overlay.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: The game no longer tricks predators and synths into thinking
they're more damaged than they actually are.
/:cl:

---------

Co-authored-by: Vero <73014819+vero5123@users.noreply.github.com>