fix gloves and map being mutually exclusive in the corpsman vendor (#6479)
# About the pull request
gloves and the map are no longer mutually exclusive in the corpsman
vendor because the gloves are now flagged as gloves instead of kit. the
map is flagged as kit.

# Explain why it's good for the game
Fixes #6476


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>

# Changelog
:cl:
fix: fixed corpsmen gloves and the map being mutually exclusive in the
corpsman vendor
/:cl:

Co-authored-by: nefarious <72267018+nefarious6th@users.noreply.github.com>