Dir assist now prioritises non-xenos (#6461)
# About the pull request
This PR makes it so when using dir assist against a xeno and a non-xeno
on the same tile the attack will prioritize the non-xeno over the xeno.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
You are more likely to want to target the marine than shove\slash the
xeno ontop of the marine.
It is annoying when John runner stands ontop of the guy you are
currently pulling and now you have to click the 3 remaining pixels
visible.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
qol: Dir-assist will now prioritize non-xenos over xenos.
/:cl: