Changes the firing sound of the M44 combat revolver (#6356)
# About the pull request

changes the firing sound of the M44 revolver to a beefier one

# Explain why it's good for the game

after the damage buff the revolver has received, the gun feels notably
more punchier but i feel as if the firing sound has stayed behind in
that regard. it sounds like a peashooter. i think that with this change
it's an objectively better sounding weapon.
 
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


old vs new comparison


https://github.com/cmss13-devs/cmss13/assets/17518895/3886d975-2207-482c-93de-f85e90f45c19



</details>


# Changelog
:cl:
soundadd: The M44 combat revolver has had its firing sound changed.
/:cl: