Makes Anti-parasitic slightly more powerful to match larva gestation increases (#6407)
# About the pull request

Larva gestation was sped up at one point or another, and that made
Anti-parasitic too slow to actually prevent gestation or remove
parasites at all
Applies a 1.5 multiplier to the potency removed from the larva timer. 

# Explain why it's good for the game

Larva gestation was sped up at one point or another, and that made
Anti-parasitic too slow to actually prevent gestation or remove
parasites at all. In its current state, Anti-Parasitic is incapable of
removing larvae at higher levels, due to it being too slow acting, with
a slight increase, it may be viable (of note, I'm not sure if the 1.5
multiplier is actually high enough to be able to counter-act and remove
larvae, and may need to be increased further)
Of note, it still does potency burn damage, meaning that if it is at
potency 8, it will do 8 burn damage to whatever the chemical is being
metabolized in.

Gives the Antiparasitic chemical property its niche back. At lower
levels it's supposed to delay the inevitable and buy time for the marine
to get to surgeons, and at higher levels it's supposed to cure larvae
from the marine. Currently, it's not powerful enough to do either to a
noticeable degree.


# Testing Photographs and Procedure

N/A. All that was done was add a 1.5 multiplier to the potency, when it
is removed from the larva timer

# Changelog
:cl:
fix: Adds a 1.5x modifier to the amount taken from larva timer based on
chem property level
fix: swaps the 1.5 multiplier for a delta_time multiplier
fix: improved the formatting of the code to have proper spacing
/:cl: