Ups protective form cost from 50 to 150 for synths (#6414)
# About the pull request
Increases the power requirement from 50 to 150 for protective form.

# Explain why it's good for the game

Synths shouldnt be able to use this ability 4 times without any cooldown
to tank the entire hive. An ability that reduces the damage you take (up
to %80 with base stats) should have some sort of a high requirement
since the "power" is already refillable instantly anyways. This will
make it so synths cant just rush into the entire hive and recover a
marine. or atleast make them think on how they do it so they arent given
free leeway due to the amount of times you can use it.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: Synthetic Protective Form now takes 150 charge instead of 50.
/:cl:

Co-authored-by: InsaneRed <prodexter31@outlook.comm>