Adds in-game bug reports without needing a GitHub account (#6392)
# About the pull request

title

# Explain why it's good for the game
Essentially, the idea is to allow for the creation of bug reports
in-game without needing to own a Github account, this can help to
incentivize issue reporting and be beneficial overall.

TODO:
- [x] Implement the Github API token to allow for anonymous bug reports
- [x] Make in-game bug reports admin approval only to prevent abuse and
to make sure only quality bug reports are approved.

Notable changes from Goonstation's bug report system is that I migrated
it to type script and it now requires admin approval.
# Changelog
Although I made alterations to meet our needs, all credit goes to Pali
for their amazing work and to the devs at Goonstation.
:cl:
add: Adds the ability for users to make bug reports in-game
ui: New bug report system ui
/:cl:

---------

Co-authored-by: DOOM <N/A>
Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>
Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>