Adds a HUD icon for crit marines for vampire lurker (#6314)
# About the pull request
Adds an HUD icon for marines that can be headbited for vampire
This was comically complicated to add
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Rather than having to spam headbite 20 times until someone is actually
unconcious, adds a neat HUD icon.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
qol: Added an hud icon for vampires that lets them know when an enemy
can be executed.
/:cl:

---------

Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>