Show facehugger limit when examining eggs/morphers (#6386)
# About the pull request

When you examine a hive core, it tells you the lesser drone limit, and
the current amount of lesser drones. This adds a similar behavior for
eggs and morphers with regards to playable facehuggers.

# Explain why it's good for the game

Helpful for ghosts to decide whether to play hugger, drone, neither, or
just for information.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl:
qol: examining eggs/morphers tells you current number of huggers and the
limit
/:cl:

---------

Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>
Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>