Remove self-extinguishing with acid spray (#6508)
# About the pull request
This PR makes it so you cannot extinguish yourself with your own spray
(this includes things that used to be possible like starting acid spray
with base praetorian and dashing to the end or extinguishing yourself
with spitter and bonus movement speed)
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Self-extinguishing discourages teamwork and makes castes that are
supposed to be vulnerable to fire able to wipe it off easily.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: You can no longer extinguish yourself with your own acid spray.
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>