Trijent Mapping Fix (#6493)
# About the pull request

Fills in a random hole in the wall.

# Explain why it's good for the game

I think this was a map oversight or an unfinished easter egg? It's just
a random hole in the wall with nothing in it - I'm mostly removing this
because Burrowers can potentially burrow in and hide there, basically
being impossible to find. It serves no purpose so I don't see why it
should stay.

# Changelog

:cl:
maptweak: Filled in a hole in the map.
/:cl: