Fix hotkeys menu crash + New/Fixed keybind support (#6565)
# About the pull request

This PR fixes a crash when using the search by key feature in the
Keybinds window when the panel is filtered, and maps a bunch of keys to
byond bindings.

Of note, there are three locations remapping keys:

https://github.com/cmss13-devs/cmss13/blob/master/tgui/packages/tgui/interfaces/KeyBinds.jsx#L14-L26

https://github.com/cmss13-devs/cmss13/blob/master/tgui/packages/common/keys.ts#L20-L39

https://github.com/cmss13-devs/cmss13/blob/master/code/_globalvars/lists/client.dm#L7-L23

Since the first two are as they are on TG, I opted to update the third
one which seems to be unique to our codebase

# Explain why it's good for the game

Fixes various keybinds (namely keypad stuff), adds a bunch of new
keybinds (media keys), and fixes a tgui crash.

Fixes: 

![image](https://github.com/cmss13-devs/cmss13/assets/76988376/8c4dd0ba-7185-4f45-a089-c8ee6edf06b7)

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![hotkeys](https://github.com/cmss13-devs/cmss13/assets/76988376/404483f5-9ea3-43e3-8c35-26e6e73d99fc)

</details>


# Changelog
:cl: Drathek
fix: Fixed a crash in the Hotkey menu when searching by key when
filtered
fix: Fixed/Added support for various keys (e.g. keypad and media keys)
/:cl:

---------

Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>