Fixes drawing weapons when knocked down (#6569)
# About the pull request
Prevents you from drawing weapons when you are knocked down
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
When you draw a weapon and are knocked down it will simply drop the
weapon since you can't hold it which is pretty annoying.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: Fixes being able to draw weapons when knocked down
/:cl:

Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>