Replaces Synth Vendor Headset (#6574)
# About the pull request

Replaces the senior command headset with the synthetic headset in the
synth equipment vendor

# Explain why it's good for the game

The synth headset is very useful as it gives synthetics a tracker for
the SLs, XO, and CO. This change will allow for survivor synthetics that
ally with the Almayer to track just as a normal synth. It also allows
shipside synths to keep the tracker if their spawn headset is
lost/stolen/melted. Also makes sense that the synth vendor gives the
synth headset


# Testing Photographs and Procedure


<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

 
:cl: Nomoresolvalou

balance: The headset in the synthetic equipment vendor has been replaced
with a synthetic headset as opposed to a senior command headset

/:cl:

Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>